First and most important part of the moderation display cleanup (will do the less important missing parts after the release).
Added Moderation::PreDisplay(), can be used to initialize values necessary in the mod display.
Removed everything that is unrelated to the actual displaying from the files in htdocs/... and moved it to PreDisplay.

Cleanup, and a lot of cosmetic changes to mod display. Include more information where necessary (like resolution in artist changes with identically named artist, links to most objects).

Some workarounds (which still look nasty) for changes in the format of the 'prevvalue' collumn, this includes guessing of missing data, so old mods look mostly the same than the new ones.

(Closes: #610, #920, #623, #853 and probably more)


git-svn-id: http://svn.musicbrainz.org/mb_server/trunk@6710 b0b80210-5d09-0410-99dd-b4bd03f891c0